+++ src/_pytask/collect.py
@@ -10,6 +10,7 @@ from pathlib import Path
 from typing import Any
 from typing import Generator
 from typing import Iterable
+from typing import TYPE_CHECKING
 
 from _pytask.collect_utils import parse_dependencies_from_task_function
 from _pytask.collect_utils import parse_products_from_task_function
@@ -18,12 +19,12 @@ from _pytask.config import IS_FILE_SYSTEM_CASE_SENSITIVE
 from _pytask.console import _get_file
 from _pytask.console import console
 from _pytask.console import create_summary_panel
-from _pytask.console import format_task_id
+from _pytask.console import format_task_name
 from _pytask.console import is_jupyter
 from _pytask.exceptions import CollectionError
 from _pytask.mark_utils import has_mark
-from _pytask.models import NodeInfo
 from _pytask.node_protocols import Node
+from _pytask.node_protocols import PTask
 from _pytask.nodes import PathNode
 from _pytask.nodes import PythonNode
 from _pytask.nodes import Task
@@ -32,12 +33,15 @@ from _pytask.outcomes import count_outcomes
 from _pytask.path import find_case_sensitive_path
 from _pytask.path import import_path
 from _pytask.report import CollectionReport
-from _pytask.session import Session
 from _pytask.shared import find_duplicates
 from _pytask.shared import reduce_node_name
 from _pytask.task_utils import task as task_decorator
 from _pytask.traceback import render_exc_info
 from rich.text import Text
+
+if TYPE_CHECKING:
+    from _pytask.session import Session
+    from _pytask.models import NodeInfo
 
 
 @hookimpl
@@ -97,7 +101,8 @@ def _collect_from_tasks(session: Session) -> None:
             )
 
         if not callable(raw_task):
-            raise ValueError("Not a function.")
+            msg = "Not a function."
+            raise ValueError(msg)
 
         if not hasattr(raw_task, "pytask_meta"):
             raw_task = task_decorator()(raw_task)  # noqa: PLW2901
@@ -126,8 +131,7 @@ def _collect_from_tasks(session: Session) -> None:
 @hookimpl
 def pytask_ignore_collect(path: Path, config: dict[str, Any]) -> bool:
     """Ignore a path during the collection."""
-    is_ignored = any(path.match(pattern) for pattern in config["ignore"])
-    return is_ignored
+    return any(path.match(pattern) for pattern in config["ignore"])
 
 
 @hookimpl
@@ -318,16 +322,16 @@ def _not_ignored_paths(
 
 
 @hookimpl(trylast=True)
-def pytask_collect_modify_tasks(tasks: list[Task]) -> None:
+def pytask_collect_modify_tasks(tasks: list[PTask]) -> None:
     """Given all tasks, assign a short uniquely identifiable name to each task."""
     id_to_short_id = _find_shortest_uniquely_identifiable_name_for_tasks(tasks)
     for task in tasks:
-        short_id = id_to_short_id[task.name]
-        task.short_name = short_id
+        if task.name in id_to_short_id and isinstance(task, Task):
+            task.display_name = id_to_short_id[task.name]
 
 
 def _find_shortest_uniquely_identifiable_name_for_tasks(
-    tasks: list[Task],
+    tasks: list[PTask],
 ) -> dict[str, str]:
     """Find the shortest uniquely identifiable name for tasks.
 
@@ -339,7 +343,7 @@ def _find_shortest_uniquely_identifiable_name_for_tasks(
     id_to_short_id = {}
 
     # Make attempt to add up to twenty parts of the path to ensure uniqueness.
-    id_to_task = {task.name: task for task in tasks}
+    id_to_task = {task.name: task for task in tasks if isinstance(task, Task)}
     for n_parts in range(1, 20):
         dupl_id_to_short_id = {
             id_: "/".join(task.path.parts[-n_parts:]) + "::" + task.base_name
@@ -361,14 +365,12 @@ def _find_shortest_uniquely_identifiable_name_for_tasks(
 
 @hookimpl
 def pytask_collect_log(
-    session: Session, reports: list[CollectionReport], tasks: list[Task]
+    session: Session, reports: list[CollectionReport], tasks: list[PTask]
 ) -> None:
     """Log collection."""
     session.collection_end = time.time()
 
     console.print(f"Collected {len(tasks)} task{'' if len(tasks) == 1 else 's'}.")
-
-    print(tasks)
 
     failed_reports = [r for r in reports if r.outcome == CollectionOutcome.FAIL]
     if failed_reports:
@@ -384,9 +386,9 @@ def pytask_collect_log(
             if report.node is None:
                 header = "Error"
             else:
-                if isinstance(report.node, Task):
-                    short_name = format_task_id(
-                        report.node, editor_url_scheme="no_link", short_name=True
+                if isinstance(report.node, PTask):
+                    short_name = format_task_name(
+                        report.node, editor_url_scheme="no_link"
                     )
                 else:
                     short_name = reduce_node_name(report.node, session.config["paths"])