+++ src/_pytask/node_protocols.py
@@ -1,12 +1,18 @@
 from __future__ import annotations
 
 from abc import abstractmethod
-from pathlib import Path
 from typing import Any
+from typing import Callable
 from typing import Protocol
 from typing import runtime_checkable
+from typing import TYPE_CHECKING
 
-from _pytask.typing import NoValue
+
+if TYPE_CHECKING:
+    from _pytask.typing import NoValue
+    from _pytask.tree_util import PyTree
+    from pathlib import Path
+    from _pytask.mark import Mark
 
 
 @runtime_checkable
@@ -45,6 +51,34 @@ class PPathNode(Node, Protocol):
     """Nodes with paths.
 
     Nodes with paths receive special handling when it comes to printing their names.
+
+    """
+
+    path: Path | NoValue
+
+
+@runtime_checkable
+class PTask(MetaNode, Protocol):
+    """Protocol for nodes."""
+
+    name: str
+    depends_on: PyTree[Node]
+    produces: PyTree[Node]
+    markers: list[Mark]
+    report_sections: list[tuple[str, str, str]]
+    attributes: dict[Any, Any]
+    function: Callable[..., Any]
+
+    def execute(self, **kwargs: Any) -> Any:
+        """Return the value of the node that will be injected into the task."""
+        ...
+
+
+@runtime_checkable
+class PTaskWithPath(PTask, Protocol):
+    """Tasks with paths.
+
+    Tasks with paths receive special handling when it comes to printing their names.
 
     """