+++ src/_pytask/nodes.py
@@ -11,9 +11,6 @@ from typing import TYPE_CHECKING
 from _pytask.node_protocols import MetaNode
 from _pytask.node_protocols import Node
 from _pytask.node_protocols import PPathNode
-from _pytask.tree_util import PyTree
-from _pytask.tree_util import tree_leaves
-from _pytask.tree_util import tree_structure
 from _pytask.typing import no_value
 from _pytask.typing import NoValue
 from attrs import define
@@ -21,6 +18,7 @@ from attrs import field
 
 
 if TYPE_CHECKING:
+    from _pytask.tree_util import PyTree
     from _pytask.mark import Mark
 
 
@@ -39,7 +37,7 @@ class Task(MetaNode):
     """The task function."""
     name: str | None = field(default=None, init=False)
     """The name of the task."""
-    short_name: str | None = field(default=None, init=False)
+    display_name: str | None = field(default=None, init=False)
     """The shortest uniquely identifiable name for task for display."""
     depends_on: PyTree[Node] = field(factory=dict)
     """A list of dependencies of task."""
@@ -47,7 +45,7 @@ class Task(MetaNode):
     """A list of products of task."""
     markers: list[Mark] = field(factory=list)
     """A list of markers attached to the task function."""
-    _report_sections: list[tuple[str, str, str]] = field(factory=list)
+    report_sections: list[tuple[str, str, str]] = field(factory=list)
     """Reports with entries for when, what, and content."""
     attributes: dict[Any, Any] = field(factory=dict)
     """A dictionary to store additional information of the task."""
@@ -60,41 +58,18 @@ class Task(MetaNode):
             else:
                 self.name = self.path.as_posix() + "::" + self.base_name
 
-        if self.short_name is None:
-            self.short_name = self.name
+        if self.display_name is None:
+            self.display_name = self.name
 
-    def state(self, hash: bool = False) -> str | None:  # noqa: A002
+    def state(self) -> str | None:
         """Return the state of the node."""
-        if hash and self.path.exists():
-            return hashlib.sha256(self.path.read_bytes()).hexdigest()
-        if not hash and self.path.exists():
+        if self.path.exists():
             return str(self.path.stat().st_mtime)
         return None
 
     def execute(self, **kwargs: Any) -> None:
         """Execute the task."""
-        out = self.function(**kwargs)
-
-        if "return" in self.produces:
-            structure_out = tree_structure(out)
-            structure_return = tree_structure(self.produces["return"])
-            # strict must be false when none is leaf.
-            if not structure_return.is_prefix(structure_out, strict=False):
-                raise ValueError(
-                    "The structure of the return annotation is not a subtree of the "
-                    "structure of the function return.\n\nFunction return: "
-                    f"{structure_out}\n\nReturn annotation: {structure_return}"
-                )
-
-            nodes = tree_leaves(self.produces["return"])
-            values = structure_return.flatten_up_to(out)
-            for node, value in zip(nodes, values):
-                node.save(value)
-
-    def add_report_section(self, when: str, key: str, content: str) -> None:
-        """Add sections which will be displayed in report like stdout or stderr."""
-        if content:
-            self._report_sections.append((when, key, content))
+        return self.function(**kwargs)
 
 
 @define(kw_only=True)
@@ -119,9 +94,11 @@ class PathNode(PPathNode):
 
         """
         if self.path is not no_value:
-            raise ValueError("'path' is already defined.")
+            msg = "'path' is already defined."
+            raise ValueError(msg)
         if not isinstance(value, Path):
-            raise TypeError("'value' must be a 'pathlib.Path'.")
+            msg = "'value' must be a 'pathlib.Path'."
+            raise TypeError(msg)
         if not self.name:
             self.name = value.as_posix()
         self.path = value
@@ -135,7 +112,8 @@ class PathNode(PPathNode):
 
         """
         if not path.is_absolute():
-            raise ValueError("Node must be instantiated from absolute path.")
+            msg = "Node must be instantiated from absolute path."
+            raise ValueError(msg)
         return cls(name=path.as_posix(), path=path)
 
     def state(self) -> str | None:
@@ -153,21 +131,22 @@ class PathNode(PPathNode):
     def load(self) -> Path:
         """Load the value."""
         if self.path is no_value:
-            raise ValueError("PathNode has no path defined.")
+            msg = "PathNode has no path defined."
+            raise ValueError(msg)
         return self.path
 
     def save(self, value: bytes | str) -> None:
         """Save strings or bytes to file."""
         if self.path is no_value:
-            raise TypeError("'path' must be a 'pathlib.Path' to store data.")
+            msg = "'path' must be a 'pathlib.Path' to store data."
+            raise TypeError(msg)
         if isinstance(value, str):
             self.path.write_text(value)
         elif isinstance(value, bytes):
             self.path.write_bytes(value)
         else:
-            raise TypeError(
-                f"'PathNode' can only save 'str' and 'bytes', not {type(value)}"
-            )
+            msg = f"'PathNode' can only save 'str' and 'bytes', not {type(value)}"
+            raise TypeError(msg)
 
 
 @define(kw_only=True)
@@ -184,7 +163,8 @@ class PythonNode(Node):
     def load(self) -> Any:
         """Load the value."""
         if self.value is no_value:
-            raise ValueError("PythonNode has no value defined.")
+            msg = "PythonNode has no value defined."
+            raise ValueError(msg)
         return self.value
 
     def save(self, value: Any) -> None:
@@ -205,7 +185,8 @@ class PythonNode(Node):
 
         """
         if self.value is not no_value:
-            raise ValueError("'value' is already defined.")
+            msg = "'value' is already defined."
+            raise ValueError(msg)
         self.value = value
 
     def state(self) -> str | None: